+++ fireREST/fmc/deployment/deployabledevice/deployment/__init__.py
@@ -7,7 +7,4 @@ class Deployment(ChildResource):
     CONTAINER_PATH = '/deployment/deployabledevices/{uuid}'
     PATH = '/deployment/deployabledevices/{container_uuid}/deployments'
     SUPPORTED_FILTERS = ['start_time', 'end_time']
-    SUPPORTED_PARAMS = []
-    IGNORE_FOR_CREATE = []
-    IGNORE_FOR_UPDATE = []
     MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_660