+++ fireREST/fmc/deployment/deployabledevice/pendingchanges/__init__.py
@@ -7,7 +7,4 @@ class PendingChanges(ChildResource):
     CONTAINER_PATH = '/deployment/deployabledevices/{uuid}'
     PATH = '/deployment/deployabledevices/{container_uuid}/pendingchanges'
     SUPPORTED_FILTERS = ['parent_entity_types', 'parent_uuid']
-    SUPPORTED_PARAMS = []
-    IGNORE_FOR_CREATE = []
-    IGNORE_FOR_UPDATE = []
     MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_660