+++ fireREST/fmc/deployment/jobhistory/__init__.py
@@ -5,5 +5,4 @@ from fireREST.fmc import Resource
 class JobHistory(Resource):
     PATH = '/deployment/jobhistories/{uuid}'
     SUPPORTED_FILTERS = ['device_uuid']
-    IGNORE_FOR_UPDATE = []
     MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_670