+++ fireREST/fmc/object/aspathlist/__init__.py
@@ -1,7 +1,10 @@
-from fireREST.defaults import API_RELEASE_660
+from fireREST.defaults import API_RELEASE_660, API_RELEASE_710
 from fireREST.fmc import Resource
 
 
 class AsPathList(Resource):
     PATH = '/object/aspathlists/{uuid}'
+    MINIMUM_VERSION_REQUIRED_CREATE = API_RELEASE_710
     MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_660
+    MINIMUM_VERSION_REQUIRED_UPDATE = API_RELEASE_710
+    MINIMUM_VERSION_REQUIRED_DELETE = API_RELEASE_710