+++ fireREST/fmc/policy/ftds2svpn/advancedsettings/__init__.py
@@ -6,9 +6,5 @@ class AdvancedSettings(ChildResource):
     CONTAINER_NAME = 'FtdS2sVpn'
     CONTAINER_PATH = '/policy/ftds2svpns/{uuid}'
     PATH = '/policy/ftds2svpns/{container_uuid}/advancedsettings/{uuid}'
-    SUPPORTED_FILTERS = []
-    SUPPORTED_PARAMS = []
-    IGNORE_FOR_CREATE = []
-    IGNORE_FOR_UPDATE = []
     MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_630
     MINIMUM_VERSION_REQUIRED_UPDATE = API_RELEASE_630