+++ fireREST/fmc/policy/prefilterpolicy/operational/hitcounts/__init__.py
@@ -7,8 +7,6 @@ class Hitcount(ChildResource):
     CONTAINER_NAME = 'PrefilterPolicy'
     CONTAINER_PATH = '/policy/prefilterpolicies/{uuid}'
     PATH = '/policy/prefilterpolicies/{container_uuid}/operational/hitcounts/{uuid}'
-    SUPPORTED_FILTERS = []
-    SUPPORTED_PARAMS = []
     MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_640
     MINIMUM_VERSION_REQUIRED_UPDATE = API_RELEASE_640
     MINIMUM_VERSION_REQUIRED_DELETE = API_RELEASE_640